fix(objects): preserve sub-degree rotation and fix rotated-image resize
Adversarial-review follow-ups: hit-testing truncated the float rotation with
int() before handing it to the view, so each new rotate-handle grab lost the
fractional angle — read it back as float now (AC-4d across repeated drags).
Resizing a rotated image used the new AABB as the image size, inflating it by
|cos|+|sin|; recover the un-rotated size by inverting the AABB↔size relation
(falling back to the current size near the singular 45°/135° case). Update the
now-float rotation type annotations on ObjectHitInfo / NativeImageInvocation.

+++ model/object_requests.py
@@ -18,7 +18,7 @@ class ObjectHitInfo:
     object_id: str
     page_num: int
     bbox: fitz.Rect
-    rotation: int = 0
+    rotation: float = 0.0
     supports_move: bool = True
     supports_delete: bool = True
     supports_rotate: bool = False
